# Improve Performance of OperandResolver#parseDouble