--- a/src/java/org/apache/poi/ss/formula/eval/OperandResolver.java
+++ b/src/java/org/apache/poi/ss/formula/eval/OperandResolver.java
@@ -41,9 +41,10 @@ public final class OperandResolver {
                  "[+-]?(" +
                    "("+Digits+"(\\.)?("+Digits+"?)("+Exp+")?)|"+
                  "(\\."+Digits+"("+Exp+")?))"+
-                 "[\\x00-\\x20]*"); 
-            
-    
+                 "[\\x00-\\x20]*");
+    private static final Pattern fpPattern = Pattern.compile(fpRegex);
+
+
     private OperandResolver() {
         // no instances of this class
     }
@@ -292,7 +293,7 @@ public final class OperandResolver {
      */
     public static Double parseDouble(String pText) {
         
-        if (Pattern.matches(fpRegex, pText))
+        if (fpPattern.matcher(pText).matches())
             try {
                 return Double.parseDouble(pText);
             } catch (NumberFormatException e) {